pyproject.toml
[project]
name = "fvc"
version = "0.1.0"
description = ""
authors = [{ name = "wooloo", email = "ayan0312000@gmail.com" }]
license = "MIT"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "click>=8.2.1",
    "insightface>=0.7.3",
    "moviepy>=2.2.1",
    "numpy>=2.2.6",
    "onnxruntime-gpu>=1.22.0",
    "opencv-python>=4.12.0.88",
    "pillow>=11.3.0",
    "torch>=2.7.1",
    "torchaudio>=2.7.1",
    "torchvision>=0.22.1",
    "tqdm>=4.67.1",
]

[project.scripts]
cli = "fvc.cli:cli"

[tool.uv.sources]
torch = [{ index = "pytorch-cu128" }]
torchaudio = [{ index = "pytorch-cu128" }]
torchvision = [{ index = "pytorch-cu128" }]

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true


[tool.hatch]
[dependency-groups]
dev = [
    "hatch>=1.14.1",
    "mypy>=1.17.0",
    "pre-commit>=4.2.0",
    "pytest>=8.4.1",
    "ruff>=0.12.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
